Baseline preprocesing pipeline implemented.
Returns a numpy array of shape (num_samples, max_num_atoms, num_features_per_atom).
For now only features are the relative x,y,z coordinates for all atoms
relative to the calculated bond center (mean coordinates of the atoms in the bond).
All molecules that are smaller than 29 atoms (max atoms) are padded with 0s.
Only other feature than the rel coordinates is an indicator if the atom row
in a sample is padding or not.